refactor: rename package to smiti18n
- Rename main directory from i18n to smiti18n
- Update all require paths to use new package name
- Update README.md examples
- Update test files to use new require paths

This change reflects the new project name after forking from i18n.lua
